Move filesystem and block workers behind interfaces
