+++ core/server/src/main/java/alluxio/worker/block/BlockWorker.java
@@ -11,215 +11,33 @@
 
 package alluxio.worker.block;
 
-import alluxio.Configuration;
-import alluxio.Constants;
-import alluxio.Sessions;
-import alluxio.exception.AlluxioException;
 import alluxio.exception.BlockAlreadyExistsException;
 import alluxio.exception.BlockDoesNotExistException;
-import alluxio.exception.ConnectionFailedException;
-import alluxio.exception.ExceptionMessage;
 import alluxio.exception.InvalidWorkerStateException;
 import alluxio.exception.WorkerOutOfSpaceException;
-import alluxio.heartbeat.HeartbeatContext;
-import alluxio.heartbeat.HeartbeatThread;
-import alluxio.thrift.AlluxioTException;
-import alluxio.thrift.BlockWorkerClientService;
-import alluxio.util.ThreadFactoryUtils;
-import alluxio.util.io.FileUtils;
-import alluxio.util.network.NetworkAddressUtils;
-import alluxio.util.network.NetworkAddressUtils.ServiceType;
 import alluxio.wire.FileInfo;
-import alluxio.wire.WorkerNetAddress;
-import alluxio.worker.AbstractWorker;
-import alluxio.worker.SessionCleaner;
-import alluxio.worker.SessionCleanupCallback;
-import alluxio.worker.WorkerContext;
-import alluxio.worker.WorkerIdRegistry;
+import alluxio.worker.Worker;
 import alluxio.worker.block.io.BlockReader;
 import alluxio.worker.block.io.BlockWriter;
 import alluxio.worker.block.meta.BlockMeta;
-import alluxio.worker.block.meta.TempBlockMeta;
-import alluxio.worker.file.FileSystemMasterClient;
-
-import com.google.common.base.Throwables;
-import org.apache.thrift.TProcessor;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.Set;
-import java.util.concurrent.Executors;
-
-import javax.annotation.concurrent.NotThreadSafe;
 
 /**
- * The class is responsible for managing all top level components of the Block Worker.
- *
- * This includes:
- *
- * Servers: {@link BlockWorkerClientServiceHandler} (RPC Server)
- *
- * Periodic Threads: {@link BlockMasterSync} (Worker to Master continuous communication)
- *
- * Logic: {@link BlockWorker} (Logic for all block related storage operations)
+ * A block worker in the Alluxio system.
  */
-@NotThreadSafe // TODO(jiri): make thread-safe (c.f. ALLUXIO-1624)
-public final class BlockWorker extends AbstractWorker {
-  private static final Logger LOG = LoggerFactory.getLogger(Constants.LOGGER_TYPE);
-
-  /** Runnable responsible for heartbeating and registration with master. */
-  private BlockMasterSync mBlockMasterSync;
-
-  /** Runnable responsible for fetching pinlist from master. */
-  private PinListSync mPinListSync;
-
-  /** Runnable responsible for clean up potential zombie sessions. */
-  private SessionCleaner mSessionCleaner;
-
-  /** Logic for handling RPC requests. */
-  private final BlockWorkerClientServiceHandler mServiceHandler;
-
-  /** Client for all block master communication. */
-  private final BlockMasterClient mBlockMasterClient;
-
-  /** Client for all file system master communication. */
-  private final FileSystemMasterClient mFileSystemMasterClient;
-
-  /** Space reserver for the block data manager. */
-  private SpaceReserver mSpaceReserver = null;
-  /** Block store delta reporter for master heartbeat. */
-  private BlockHeartbeatReporter mHeartbeatReporter;
-  /** Metrics reporter that listens on block events and increases metrics counters. */
-  private BlockMetricsReporter mMetricsReporter;
-  /** Session metadata, used to keep track of session heartbeats. */
-  private Sessions mSessions;
-  /** Block Store manager. */
-  private BlockStore mBlockStore;
-
+public interface BlockWorker extends Worker {
   /**
    * @return the worker data service bind host
    */
-  public BlockStore getBlockStore() {
-    return mBlockStore;
-  }
+  BlockStore getBlockStore();
 
   /**
    * @return the worker service handler
    */
-  public BlockWorkerClientServiceHandler getWorkerServiceHandler() {
-    return mServiceHandler;
-  }
-
-  /**
-   * Creates a new instance of {@link BlockWorker}.
-   *
-   * @throws IOException for other exceptions
-   */
-  public BlockWorker() throws IOException {
-    super(Executors.newFixedThreadPool(4,
-        ThreadFactoryUtils.build("block-worker-heartbeat-%d", true)));
-    // Setup BlockMasterClient
-    mBlockMasterClient =
-        new BlockMasterClient(NetworkAddressUtils.getConnectAddress(ServiceType.MASTER_RPC));
-
-    mFileSystemMasterClient =
-        new FileSystemMasterClient(NetworkAddressUtils.getConnectAddress(ServiceType.MASTER_RPC));
-
-    // Setup RPC ServerHandler
-    mServiceHandler = new BlockWorkerClientServiceHandler(this);
-
-    // Setup BlockHeartbeatReporter
-    mHeartbeatReporter = new BlockHeartbeatReporter();
-    // Setup BlockMetricsReporter
-    mMetricsReporter = new BlockMetricsReporter(WorkerContext.getWorkerSource());
-    // Setup Sessions
-    mSessions = new Sessions();
-    // Setup the BlockStore
-    mBlockStore = new TieredBlockStore();
-
-    // Register the heartbeat reporter so it can record block store changes
-    mBlockStore.registerBlockStoreEventListener(mHeartbeatReporter);
-    mBlockStore.registerBlockStoreEventListener(mMetricsReporter);
-  }
-
-  @Override
-  public Map<String, TProcessor> getServices() {
-    Map<String, TProcessor> services = new HashMap<>();
-    services.put(Constants.BLOCK_WORKER_CLIENT_SERVICE_NAME,
-        new BlockWorkerClientService.Processor<>(getWorkerServiceHandler()));
-    return services;
-  }
-
-  /**
-   * Runs the block worker. The thread must be called after all services (e.g., web, dataserver)
-   * started.
-   *
-   * @throws IOException if a non-Alluxio related exception occurs
-   */
-  @Override
-  public void start() throws IOException {
-    WorkerNetAddress netAddress;
-    try {
-      netAddress = WorkerContext.getNetAddress();
-      WorkerIdRegistry.registerWithBlockMaster(mBlockMasterClient, netAddress);
-    } catch (ConnectionFailedException e) {
-      LOG.error("Failed to get a worker id from block master", e);
-      throw Throwables.propagate(e);
-    }
-
-    // Setup BlockMasterSync
-    mBlockMasterSync =
-        new BlockMasterSync(this, netAddress, mBlockMasterClient);
-
-    // Setup PinListSyncer
-    mPinListSync = new PinListSync(this, mFileSystemMasterClient);
-
-    // Setup session cleaner
-    setupSessionCleaner();
-
-    // Setup space reserver
-    if (Configuration.getBoolean(Constants.WORKER_TIERED_STORE_RESERVER_ENABLED)) {
-      mSpaceReserver = new SpaceReserver(this);
-    }
-
-    getExecutorService()
-        .submit(new HeartbeatThread(HeartbeatContext.WORKER_BLOCK_SYNC, mBlockMasterSync,
-            Configuration.getInt(Constants.WORKER_BLOCK_HEARTBEAT_INTERVAL_MS)));
-
-    // Start the pinlist syncer to perform the periodical fetching
-    getExecutorService()
-        .submit(new HeartbeatThread(HeartbeatContext.WORKER_PIN_LIST_SYNC, mPinListSync,
-            Configuration.getInt(Constants.WORKER_BLOCK_HEARTBEAT_INTERVAL_MS)));
-
-    // Start the session cleanup checker to perform the periodical checking
-    getExecutorService().submit(mSessionCleaner);
-
-    // Start the space reserver
-    if (mSpaceReserver != null) {
-      getExecutorService().submit(mSpaceReserver);
-    }
-  }
-
-  /**
-   * Stops the block worker. This method should only be called to terminate the worker.
-   *
-   * @throws IOException if the data server fails to close
-   */
-  @Override
-  public void stop() throws IOException {
-    mSessionCleaner.stop();
-    mBlockMasterClient.close();
-    if (mSpaceReserver != null) {
-      mSpaceReserver.stop();
-    }
-    mFileSystemMasterClient.close();
-    // Use shutdownNow because HeartbeatThreads never finish until they are interrupted
-    getExecutorService().shutdownNow();
-  }
+  BlockWorkerClientServiceHandler getWorkerServiceHandler();
 
   /**
    * Aborts the temporary block created by the session.
@@ -231,10 +49,8 @@ public final class BlockWorker extends AbstractWorker {
    * @throws InvalidWorkerStateException if blockId does not belong to sessionId
    * @throws IOException if temporary block cannot be deleted
    */
-  public void abortBlock(long sessionId, long blockId) throws BlockAlreadyExistsException,
-      BlockDoesNotExistException, InvalidWorkerStateException, IOException {
-    mBlockStore.abortBlock(sessionId, blockId);
-  }
+  void abortBlock(long sessionId, long blockId) throws BlockAlreadyExistsException,
+      BlockDoesNotExistException, InvalidWorkerStateException, IOException;
 
   /**
    * Access the block for a given session. This should be called to update the evictor when
@@ -245,9 +61,7 @@ public final class BlockWorker extends AbstractWorker {
    * @throws BlockDoesNotExistException this exception is not thrown in the tiered block store
    *         implementation
    */
-  public void accessBlock(long sessionId, long blockId) throws BlockDoesNotExistException {
-    mBlockStore.accessBlock(sessionId, blockId);
-  }
+  void accessBlock(long sessionId, long blockId) throws BlockDoesNotExistException;
 
   /**
    * Commits a block to Alluxio managed space. The block must be temporary. The block is persisted
@@ -262,35 +76,9 @@ public final class BlockWorker extends AbstractWorker {
    * @throws IOException if the block cannot be moved from temporary path to committed path
    * @throws WorkerOutOfSpaceException if there is no more space left to hold the block
    */
-  public void commitBlock(long sessionId, long blockId)
+  void commitBlock(long sessionId, long blockId)
       throws BlockAlreadyExistsException, BlockDoesNotExistException, InvalidWorkerStateException,
-      IOException, WorkerOutOfSpaceException {
-    // NOTE: this may be invoked multiple times due to retry on client side.
-    // TODO(binfan): find a better way to handle retry logic
-    try {
-      mBlockStore.commitBlock(sessionId, blockId);
-    } catch (BlockAlreadyExistsException e) {
-      LOG.debug("Block {} has been in block store, this could be a retry due to master-side RPC "
-          + "failure, therefore ignore the exception", blockId, e);
-    }
-
-    // TODO(calvin): Reconsider how to do this without heavy locking.
-    // Block successfully committed, update master with new block metadata
-    Long lockId = mBlockStore.lockBlock(sessionId, blockId);
-    try {
-      BlockMeta meta = mBlockStore.getBlockMeta(sessionId, blockId, lockId);
-      BlockStoreLocation loc = meta.getBlockLocation();
-      Long length = meta.getBlockSize();
-      BlockStoreMeta storeMeta = mBlockStore.getBlockStoreMeta();
-      Long bytesUsedOnTier = storeMeta.getUsedBytesOnTiers().get(loc.tierAlias());
-      mBlockMasterClient.commitBlock(WorkerIdRegistry.getWorkerId(), bytesUsedOnTier,
-          loc.tierAlias(), blockId, length);
-    } catch (AlluxioTException | IOException | ConnectionFailedException e) {
-      throw new IOException(ExceptionMessage.FAILED_COMMIT_BLOCK_TO_MASTER.getMessage(blockId), e);
-    } finally {
-      mBlockStore.unlockBlock(lockId);
-    }
-  }
+      IOException, WorkerOutOfSpaceException;
 
   /**
    * Creates a block in Alluxio managed space. The block will be temporary until it is committed.
@@ -307,14 +95,8 @@ public final class BlockWorker extends AbstractWorker {
    * @throws WorkerOutOfSpaceException if this Store has no more space than the initialBlockSize
    * @throws IOException if blocks in eviction plan fail to be moved or deleted
    */
-  public String createBlock(long sessionId, long blockId, String tierAlias, long initialBytes)
-      throws BlockAlreadyExistsException, WorkerOutOfSpaceException, IOException {
-    BlockStoreLocation loc = BlockStoreLocation.anyDirInTier(tierAlias);
-    TempBlockMeta createdBlock = mBlockStore.createBlockMeta(sessionId, blockId, loc, initialBytes);
-    String blockPath = createdBlock.getPath();
-    createBlockFile(blockPath);
-    return blockPath;
-  }
+  String createBlock(long sessionId, long blockId, String tierAlias, long initialBytes)
+      throws BlockAlreadyExistsException, WorkerOutOfSpaceException, IOException;
 
   /**
    * Creates a block. This method is only called from a data server.
@@ -330,12 +112,8 @@ public final class BlockWorker extends AbstractWorker {
    * @throws WorkerOutOfSpaceException if this Store has no more space than the initialBlockSize
    * @throws IOException if blocks in eviction plan fail to be moved or deleted
    */
-  public void createBlockRemote(long sessionId, long blockId, String tierAlias, long initialBytes)
-      throws BlockAlreadyExistsException, WorkerOutOfSpaceException, IOException {
-    BlockStoreLocation loc = BlockStoreLocation.anyDirInTier(tierAlias);
-    TempBlockMeta createdBlock = mBlockStore.createBlockMeta(sessionId, blockId, loc, initialBytes);
-    createBlockFile(createdBlock.getPath());
-  }
+  void createBlockRemote(long sessionId, long blockId, String tierAlias, long initialBytes)
+      throws BlockAlreadyExistsException, WorkerOutOfSpaceException, IOException;
 
   /**
    * Frees space to make a specific amount of bytes available in the tier.
@@ -349,12 +127,9 @@ public final class BlockWorker extends AbstractWorker {
    * @throws BlockAlreadyExistsException if blocks to move already exists in destination location
    * @throws InvalidWorkerStateException if blocks to move/evict is uncommitted
    */
-  public void freeSpace(long sessionId, long availableBytes, String tierAlias)
+  void freeSpace(long sessionId, long availableBytes, String tierAlias)
       throws WorkerOutOfSpaceException, BlockDoesNotExistException, IOException,
-      BlockAlreadyExistsException, InvalidWorkerStateException {
-    BlockStoreLocation location = BlockStoreLocation.anyDirInTier(tierAlias);
-    mBlockStore.freeSpace(sessionId, availableBytes, location);
-  }
+      BlockAlreadyExistsException, InvalidWorkerStateException;
 
   /**
    * Opens a {@link BlockWriter} for an existing temporary block. This method is only called from a
@@ -369,10 +144,8 @@ public final class BlockWorker extends AbstractWorker {
    * @throws BlockDoesNotExistException if the block cannot be found
    * @throws IOException if block cannot be created
    */
-  public BlockWriter getTempBlockWriterRemote(long sessionId, long blockId)
-      throws BlockDoesNotExistException, IOException {
-    return mBlockStore.getBlockWriter(sessionId, blockId);
-  }
+  BlockWriter getTempBlockWriterRemote(long sessionId, long blockId)
+      throws BlockDoesNotExistException, IOException;
 
   /**
    * Gets a report for the periodic heartbeat to master. Contains the blocks added since the last
@@ -380,9 +153,7 @@ public final class BlockWorker extends AbstractWorker {
    *
    * @return a block heartbeat report
    */
-  public BlockHeartbeatReport getReport() {
-    return mHeartbeatReporter.generateReport();
-  }
+  BlockHeartbeatReport getReport();
 
   /**
    * Gets the metadata for the entire block store. Contains the block mapping per storage dir and
@@ -390,9 +161,7 @@ public final class BlockWorker extends AbstractWorker {
    *
    * @return the block store metadata
    */
-  public BlockStoreMeta getStoreMeta() {
-    return mBlockStore.getBlockStoreMeta();
-  }
+  BlockStoreMeta getStoreMeta();
 
   /**
    * Similar as {@link BlockWorker#getStoreMeta} except that this also contains full blockId
@@ -400,9 +169,7 @@ public final class BlockWorker extends AbstractWorker {
    *
    * @return the full block store metadata
    */
-  public BlockStoreMeta getStoreMetaFull() {
-    return mBlockStore.getBlockStoreMetaFull();
-  }
+  BlockStoreMeta getStoreMetaFull();
 
   /**
    * Gets the metadata of a block given its blockId or throws IOException. This method does not
@@ -412,9 +179,7 @@ public final class BlockWorker extends AbstractWorker {
    * @return metadata of the block
    * @throws BlockDoesNotExistException if no {@link BlockMeta} for this blockId is found
    */
-  public BlockMeta getVolatileBlockMeta(long blockId) throws BlockDoesNotExistException {
-    return mBlockStore.getVolatileBlockMeta(blockId);
-  }
+  BlockMeta getVolatileBlockMeta(long blockId) throws BlockDoesNotExistException;
 
   /**
    * Gets the meta data of a specific block from local storage.
@@ -431,10 +196,8 @@ public final class BlockWorker extends AbstractWorker {
    * @throws InvalidWorkerStateException if session id or block id is not the same as that in the
    *         LockRecord of lockId
    */
-  public BlockMeta getBlockMeta(long sessionId, long blockId, long lockId)
-      throws BlockDoesNotExistException, InvalidWorkerStateException {
-    return mBlockStore.getBlockMeta(sessionId, blockId, lockId);
-  }
+  BlockMeta getBlockMeta(long sessionId, long blockId, long lockId)
+      throws BlockDoesNotExistException, InvalidWorkerStateException;
 
   /**
    * Checks if the storage has a given block.
@@ -442,9 +205,7 @@ public final class BlockWorker extends AbstractWorker {
    * @param blockId the block id
    * @return true if the block is contained, false otherwise
    */
-  public boolean hasBlockMeta(long blockId) {
-    return mBlockStore.hasBlockMeta(blockId);
-  }
+  boolean hasBlockMeta(long blockId);
 
   /**
    * Obtains a read lock the block.
@@ -454,9 +215,7 @@ public final class BlockWorker extends AbstractWorker {
    * @return the lock id that uniquely identifies the lock obtained
    * @throws BlockDoesNotExistException if blockId cannot be found, for example, evicted already
    */
-  public long lockBlock(long sessionId, long blockId) throws BlockDoesNotExistException {
-    return mBlockStore.lockBlock(sessionId, blockId);
-  }
+  long lockBlock(long sessionId, long blockId) throws BlockDoesNotExistException;
 
   /**
    * Moves a block from its current location to a target location, currently only tier level moves
@@ -474,24 +233,9 @@ public final class BlockWorker extends AbstractWorker {
    *         block
    * @throws IOException if block cannot be moved from current location to newLocation
    */
-  public void moveBlock(long sessionId, long blockId, String tierAlias)
+  void moveBlock(long sessionId, long blockId, String tierAlias)
       throws BlockDoesNotExistException, BlockAlreadyExistsException, InvalidWorkerStateException,
-      WorkerOutOfSpaceException, IOException {
-    // TODO(calvin): Move this logic into BlockStore#moveBlockInternal if possible
-    // Because the move operation is expensive, we first check if the operation is necessary
-    BlockStoreLocation dst = BlockStoreLocation.anyDirInTier(tierAlias);
-    long lockId = mBlockStore.lockBlock(sessionId, blockId);
-    try {
-      BlockMeta meta = mBlockStore.getBlockMeta(sessionId, blockId, lockId);
-      if (meta.getBlockLocation().belongsTo(dst)) {
-        return;
-      }
-    } finally {
-      mBlockStore.unlockBlock(lockId);
-    }
-    // Execute the block move if necessary
-    mBlockStore.moveBlock(sessionId, blockId, dst);
-  }
+      WorkerOutOfSpaceException, IOException;
 
   /**
    * Gets the path to the block file in local storage. The block must be a permanent block, and the
@@ -506,11 +250,8 @@ public final class BlockWorker extends AbstractWorker {
    * @throws InvalidWorkerStateException if sessionId or blockId is not the same as that in the
    *         LockRecord of lockId
    */
-  public String readBlock(long sessionId, long blockId, long lockId)
-      throws BlockDoesNotExistException, InvalidWorkerStateException {
-    BlockMeta meta = mBlockStore.getBlockMeta(sessionId, blockId, lockId);
-    return meta.getPath();
-  }
+  String readBlock(long sessionId, long blockId, long lockId)
+      throws BlockDoesNotExistException, InvalidWorkerStateException;
 
   /**
    * Gets the block reader for the block. This method is only called by a data server.
@@ -524,10 +265,8 @@ public final class BlockWorker extends AbstractWorker {
    *         LockRecord of lockId
    * @throws IOException if block cannot be read
    */
-  public BlockReader readBlockRemote(long sessionId, long blockId, long lockId)
-      throws BlockDoesNotExistException, InvalidWorkerStateException, IOException {
-    return mBlockStore.getBlockReader(sessionId, blockId, lockId);
-  }
+  BlockReader readBlockRemote(long sessionId, long blockId, long lockId)
+      throws BlockDoesNotExistException, InvalidWorkerStateException, IOException;
 
   /**
    * Frees a block from Alluxio managed space.
@@ -538,10 +277,8 @@ public final class BlockWorker extends AbstractWorker {
    * @throws BlockDoesNotExistException if block cannot be found
    * @throws IOException if block cannot be removed from current path
    */
-  public void removeBlock(long sessionId, long blockId)
-      throws InvalidWorkerStateException, BlockDoesNotExistException, IOException {
-    mBlockStore.removeBlock(sessionId, blockId);
-  }
+  void removeBlock(long sessionId, long blockId)
+      throws InvalidWorkerStateException, BlockDoesNotExistException, IOException;
 
   /**
    * Request an amount of space for a block in its storage directory. The block must be a temporary
@@ -556,10 +293,8 @@ public final class BlockWorker extends AbstractWorker {
    * @throws IOException if blocks in {@link alluxio.worker.block.evictor.EvictionPlan} fail to be
    *         moved or deleted on file system
    */
-  public void requestSpace(long sessionId, long blockId, long additionalBytes)
-      throws BlockDoesNotExistException, WorkerOutOfSpaceException, IOException {
-    mBlockStore.requestSpace(sessionId, blockId, additionalBytes);
-  }
+  void requestSpace(long sessionId, long blockId, long additionalBytes)
+      throws BlockDoesNotExistException, WorkerOutOfSpaceException, IOException;
 
   /**
    * Releases the lock with the specified lock id.
@@ -567,9 +302,7 @@ public final class BlockWorker extends AbstractWorker {
    * @param lockId the id of the lock to release
    * @throws BlockDoesNotExistException if lock id cannot be found
    */
-  public void unlockBlock(long lockId) throws BlockDoesNotExistException {
-    mBlockStore.unlockBlock(lockId);
-  }
+  void unlockBlock(long lockId) throws BlockDoesNotExistException;
 
   /**
    * Releases the lock with the specified session and block id.
@@ -579,9 +312,7 @@ public final class BlockWorker extends AbstractWorker {
    * @throws BlockDoesNotExistException if block id cannot be found
    */
   // TODO(calvin): Remove when lock and reads are separate operations.
-  public void unlockBlock(long sessionId, long blockId) throws BlockDoesNotExistException {
-    mBlockStore.unlockBlock(sessionId, blockId);
-  }
+  void unlockBlock(long sessionId, long blockId) throws BlockDoesNotExistException;
 
   /**
    * Handles the heartbeat from a client.
@@ -589,37 +320,14 @@ public final class BlockWorker extends AbstractWorker {
    * @param sessionId the id of the client
    * @param metrics the set of metrics the client has gathered since the last heartbeat
    */
-  public void sessionHeartbeat(long sessionId, List<Long> metrics) {
-    mSessions.sessionHeartbeat(sessionId);
-    mMetricsReporter.updateClientMetrics(metrics);
-  }
-
-  /**
-   * Sets up the session cleaner thread. This logic is isolated for testing the session cleaner.
-   */
-  private void setupSessionCleaner() {
-    mSessionCleaner = new SessionCleaner(new SessionCleanupCallback() {
-      /**
-       * Cleans up after sessions, to prevent zombie sessions holding local resources.
-       */
-      @Override
-      public void cleanupSessions() {
-        for (long session : mSessions.getTimedOutSessions()) {
-          mSessions.removeSession(session);
-          mBlockStore.cleanupSession(session);
-        }
-      }
-    });
-  }
+  void sessionHeartbeat(long sessionId, List<Long> metrics);
 
   /**
    * Sets the pinlist for the underlying block store. Typically called by {@link PinListSync}.
    *
    * @param pinnedInodes a set of pinned inodes
    */
-  public void updatePinList(Set<Long> pinnedInodes) {
-    mBlockStore.updatePinnedInodes(pinnedInodes);
-  }
+  void updatePinList(Set<Long> pinnedInodes);
 
   /**
    * Gets the file information.
@@ -628,27 +336,5 @@ public final class BlockWorker extends AbstractWorker {
    * @return the file info
    * @throws IOException if an I/O error occurs
    */
-  public FileInfo getFileInfo(long fileId) throws IOException {
-    try {
-      return mFileSystemMasterClient.getFileInfo(fileId);
-    } catch (AlluxioException e) {
-      throw new IOException(e);
-    }
-  }
-
-  /**
-   * Creates a file to represent a block denoted by the given block path. This file will be owned
-   * by the Alluxio worker but have 777 permissions so processes under users different from the
-   * user that launched the Alluxio worker can read and write to the file. The tiered storage
-   * directory has the sticky bit so only the worker user can delete or rename files it creates.
-   *
-   * @param blockPath the block path to create
-   * @throws IOException if the file cannot be created in the tiered storage folder
-   */
-  private void createBlockFile(String blockPath) throws IOException {
-    FileUtils.createBlockPath(blockPath);
-    FileUtils.createFile(blockPath);
-    FileUtils.changeLocalFileToFullPermission(blockPath);
-    LOG.debug("Created new file block, block path: {}", blockPath);
-  }
+  FileInfo getFileInfo(long fileId) throws IOException;
 }
