+++ core/server/src/test/java/alluxio/worker/block/SpaceReserverTest.java
@@ -79,7 +79,7 @@ public class SpaceReserverTest {
         DataServer.Factory.create(Mockito.<InetSocketAddress>any(), Mockito.<AlluxioWorker>any()))
         .thenReturn(Mockito.mock(DataServer.class));
 
-    BlockWorker blockWorker = new BlockWorker();
+    BlockWorker blockWorker = new DefaultBlockWorker();
     mBlockStore = blockWorker.getBlockStore();
     String reserveRatioProp =
         String.format(Constants.WORKER_TIERED_STORE_LEVEL_RESERVED_RATIO_FORMAT, 0);
